[INJIMOB-3497]: Add Test utils for digest multibase generation
